# proto/gen/apidocs.swagger.yaml
swagger: "2.0"
info:
  title: api/v1/activity_service.proto
  version: version not set
...
consumes:
  - application/json
produces:
  - application/json
paths:
  /api/v1/ai/generate:
    post:
      summary: GenerateContent generates content from the AI.
      operationId: AIService_GenerateContent
      responses:
        "200":
          description: A successful response.(streaming responses)
          schema:
            type: object
            properties:
              result:
                $ref: '#/definitions/v1GenerateContentResponse'
              error:
                $ref: '#/definitions/googlerpcStatus'
            title: Stream result of v1GenerateContentResponse
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/googlerpcStatus'
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/v1GenerateContentRequest'
      tags:
        - AIService
  /api/v1/ai/sessions:
    get:
      summary: ListChatSessions lists chat sessions for the current user.
      operationId: AIService_ListChatSessions
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/v1ListChatSessionsResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/googlerpcStatus'
      parameters:
        - name: pageSize
          description: The maximum number of sessions to return.
          in: query
          required: false
          type: integer
          format: int32
        - name: pageToken
          description: A page token from a previous call.
          in: query
          required: false
          type: string
      tags:
        - AIService
  /api/v1/ai/sessions/{uid}:
    get:
      summary: GetChatSession gets a specific chat session.
      operationId: AIService_GetChatSession
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/v1ChatSession'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/googlerpcStatus'
      parameters:
        - name: uid
          description: The UID of the session to retrieve.
          in: path
          required: true
          type: string
      tags:
        - AIService
    delete:
      summary: DeleteChatSession deletes a chat session.
      operationId: AIService_DeleteChatSession
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/v1DeleteChatSessionResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/googlerpcStatus'
      parameters:
        - name: uid
          description: The UID of the session to delete.
          in: path
          required: true
          type: string
      tags:
        - AIService
    patch:
      summary: UpdateChatSession updates a chat session.
      operationId: AIService_UpdateChatSession
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/v1UpdateChatSessionResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/googlerpcStatus'
      parameters:
        - name: uid
          description: The UID of the session to update.
          in: path
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/AIServiceUpdateChatSessionBody'
      tags:
        - AIService
...
definitions:
  AIServiceUpdateChatSessionBody:
    type: object
    properties:
      title:
        type: string
        description: The new title of the session.
...
  googlerpcStatus:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      details:
        type: array
        items:
          type: object
          $ref: '#/definitions/protobufAny'
  protobufAny:
    type: object
    properties:
      '@type':
        type: string
        description: |-
          A URL/resource name that uniquely identifies the type of the serialized
          protocol buffer message. This string must contain at least
          one "/" character. The last segment of the URL's path must represent
          the fully qualified name of the type (as in
          `path/google.protobuf.Duration`). The name should be in a canonical form
          (e.g., leading "." is not accepted).

          In practice, teams usually precompile into the binary all types that they
          expect it to use in the context of Any. However, for URLs which use the
          scheme `http`, `https`, or no scheme, one can optionally set up a type
          server that maps type URLs to message definitions as follows:

          * If no scheme is provided, `https` is assumed.
          * An HTTP GET on the URL must yield a [google.protobuf.Type][]
            value in binary format, or produce an error.
          * Applications are allowed to cache lookup results based on the
            URL, or have them precompiled into a binary to avoid any
            lookup. Therefore, binary compatibility needs to be preserved
            on changes to types. (Use versioned type names to manage
            breaking changes.)

          Note: this functionality is not currently available in the official
          protobuf release, and it is not used for type URLs beginning with
          type.googleapis.com. As of May 2023, there are no widely used type server
          implementations and no plans to implement one.

          Schemes other than `http`, `https` (or the empty scheme) might be
          used with implementation specific semantics.
    additionalProperties: {}
    description: |-
      `Any` contains an arbitrary serialized protocol buffer message along with a
      URL that describes the type of the serialized message.

      Protobuf library provides support to pack/unpack Any values in the form
      of utility functions or additional generated methods of the Any type.

      Example 1: Pack and unpack a message in C++.

          Foo foo = ...;
          Any any;
          any.PackFrom(foo);
          ...
          if (any.UnpackTo(&foo)) {
            ...
          }

      Example 2: Pack and unpack a message in Java.

          Foo foo = ...;
          Any any = Any.pack(foo);
          ...
          if (any.is(Foo.class)) {
            foo = any.unpack(Foo.class);
          }
          // or ...
          if (any.isSameTypeAs(Foo.getDefaultInstance())) {
            foo = any.unpack(Foo.getDefaultInstance());
          }

       Example 3: Pack and unpack a message in Python.

          foo = Foo(...)
          any = Any()
          any.Pack(foo)
          ...
          if any.Is(Foo.DESCRIPTOR):
            any.Unpack(foo)
            ...

       Example 4: Pack and unpack a message in Go

           foo := &pb.Foo{...}
           any, err := anypb.New(foo)
           if err != nil {
             ...
           }
           ...
           foo := &pb.Foo{}
           if err := any.UnmarshalTo(foo); err != nil {
             ...
           }

      The pack methods provided by protobuf library will by default use
      'type.googleapis.com/full.type.name' as the type URL and the unpack
      methods only use the fully qualified type name after the last '/'
      in the type URL, for example "foo.bar.com/x/y.z" will yield type
      name "y.z".

      JSON
      ====
      The JSON representation of an `Any` value uses the regular
      representation of the deserialized, embedded message, with an
      additional field `@type` which contains the type URL. Example:

          package google.profile;
          message Person {
            string first_name = 1;
            string last_name = 2;
          }

          {
            "@type": "type.googleapis.com/google.profile.Person",
            "firstName": <string>,
            "lastName": <string>
          }

      If the embedded message type is well-known and has a custom JSON
      representation, that representation will be embedded adding a field
      `value` which holds the custom JSON in addition to the `@type`
      field. Example (for message [google.protobuf.Duration][]):

          {
            "@type": "type.googleapis.com/google.protobuf.Duration",
            "value": "1.212s"
          }
...
  v1ChatMessage:
    type: object
    properties:
      role:
        type: string
        description: The role of the message.
      content:
        type: string
        description: The content of the message.
      createdTime:
        type: string
        format: date-time
        description: The timestamp when the message was created.
  v1ChatSession:
    type: object
    properties:
      uid:
        type: string
        description: The unique identifier for the session.
      title:
        type: string
        description: The title of the session.
      createdTime:
        type: string
        format: date-time
        description: The timestamp when the session was created.
      updatedTime:
        type: string
        format: date-time
        description: The timestamp when the session was last updated.
      status:
        type: string
        description: The status of the session.
      messages:
        type: array
        items:
          type: object
          $ref: '#/definitions/v1ChatMessage'
        description: The messages in the session.
...
  v1DeleteChatSessionResponse:
    type: object
...
  v1GenerateContentRequest:
    type: object
    properties:
      messages:
        type: array
        items:
          type: object
          $ref: '#/definitions/v1ChatMessage'
        description: The content of the request.
      sessionUid:
        type: string
        description: Optional session UID to save the conversation.
  v1GenerateContentResponse:
    type: object
    properties:
      content:
        type: string
        description: The content of the response.
...
  v1ListChatSessionsResponse:
    type: object
    properties:
      sessions:
        type: array
        items:
          type: object
          $ref: '#/definitions/v1ChatSession'
        description: The list of sessions.
      nextPageToken:
        type: string
        description: A token to retrieve the next page of results.
...
  v1UpdateChatSessionResponse:
    type: object
    properties:
      session:
        $ref: '#/definitions/v1ChatSession'
        description: The updated session.
...